Изменения по замечанию v2

+++ src/main/java/ru/yandex/mymarketappskeleton/mapper/ItemMapper.java
@@ -13,7 +13,7 @@ public class ItemMapper {
                 .title(item.getTitle())
                 .description(item.getDescription())
                 .imgPath(item.getImgPath())
-                .price(item.getPrice().doubleValue())
+                .price(item.getPrice())
                 .build();
 
     }
